+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [2.1.0] - 2023-01-13
+
+### Added
+
+-   Add i2c presence check at gpio pca9554 init. That way an error message is 
+    logged, if pca9554 chip is not present on the i2c bus.
+
+### Changed
+
+-   Make log messages lower-case, for consistency.
+
 ## [2.0.0] - 2023-01-06
 
 ### Changed
@@ -26,7 +37,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
     `nrf-sdk v2.1.0.`.
 -   Sample which shows basic use and configuration.
 
-[Unreleased]: https://github.com/IRNAS/irnas-pca9554-driver/compare/v2.0.0...HEAD
+[Unreleased]: https://github.com/IRNAS/irnas-pca9554-driver/compare/v2.1.0...HEAD
+
+[2.1.0]: https://github.com/IRNAS/irnas-pca9554-driver/compare/v2.0.0...v2.1.0
 
 [2.0.0]: https://github.com/IRNAS/irnas-pca9554-driver/compare/v1.0.0...v2.0.0